api/requirements.txt
django
djangorestframework
pandas
imgkit
requests
pillow
pymongo
